requirements.txt
celery==4.4.7
coreapi==2.3.3
cryptography==3.0
six==1.15.0
Django==3.1
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-cors-headers==3.4.0
django-oauth-toolkit==1.3.2
djangorestframework==3.11.1
django-timezone-field==4.0
Pillow==7.2.0
redis==3.5.3
requests==2.24.0
selenium==3.141.0
uWSGI==2.0.19.1
mysqlclient==2.0.1
pytest-django==3.10.0
pytest-cov==2.10.1
pytest-html==2.1.1
